+++ setup/nodegoat_home.various.sql
@@ -2,5 +2,5 @@ INSERT INTO `def_custom_content` (`id`, `name`, `body`, `style`, `script`, `desc
 (14, 'Footer', '<a class=\"nodegoat-logo\" href=\"https://nodegoat.net\"></a>\r\n<p><span>nodegoat is developed by</span><a href=\"https://lab1100.com\" target=\"_blank\" class=\"LAB1100\"></a></p>\r\n<p>version [S][version]</p>', 'default', '', ''),
 (16, 'Logo', '<a class=\"nodegoat-logo\" href=\"/\"></a>', 'default', '', '');
 
-INSERT INTO `def_nodegoat_details` (`unique_row`, `limit_view`) VALUES
-(1, 150000);
+INSERT INTO `def_nodegoat_details` (`unique_row`, `limit_view`, `limit_import`) VALUES
+(1, 150000, 50000);