+++ CHANGELOG.md
@@ -6,6 +6,10 @@ To get the version for your current 1100CC installation, view `./APP/CORE/CMS/in
 
 Initial release.
 
-## VERSION x.x
+## VERSION 7.2
 
-Nothing yet.
+* Fixed an error when toggling for default value in data_model.php
+* Improved user interaction to use weighted data in the visualisations based on Conditions and MapGeo/MapSocial/MapTimeline.js
+* Renamed Analysis metric 'Count' to 'Degree Centrality' and added the weight option.
+* Fixed/improved the combination of touch an mouse interaction in MapScroller.js and MapSocial.js.
+* Added the option to limit the amount of rows from one CSV when importing data. This can be adjusted in the CMS under nodegoat Details.